Updated seeds, wrote query for fetching warranties from db

+++ express-back-end/db/schema/03_entries.sql
@@ -5,7 +5,7 @@ CREATE TABLE entries (
   item_id INTEGER REFERENCES items(id),
   type VARCHAR(255) NOT NULL,
   duration_in_months INT NOT NULL,
-  start_date INT NOT NULL,
+  start_date BIGINT NOT NULL,
   sms BOOLEAN NOT NULL,
   email BOOLEAN NOT NULL,
   days_prior INT NOT NULL
